requirement.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
lxml==4.4.1
mysql-connector==2.2.9
PyMySQL==0.9.3
qqwry-py3==1.0.8
requests==2.22.0
SQLAlchemy==1.3.10
urllib3==1.25.6
